Use WeakConcurrent in AttributeKeys
