+++ dd-java-agent/agent-tooling/src/main/java/datadog/trace/agent/tooling/AgentTooling.java
@@ -22,7 +22,7 @@ public class AgentTooling {
 
   private static void registerWeakMapProvider() {
     if (!WeakMap.Provider.isProviderRegistered()) {
-      WeakMap.Provider.registerIfAbsent(new WeakMapSuppliers.WeakConcurrent(new Cleaner()));
+      WeakMap.Provider.registerIfAbsent(new WeakMapSuppliers.WeakConcurrent());
       //    WeakMap.Provider.registerIfAbsent(new WeakMapSuppliers.WeakConcurrent.Inline());
       //    WeakMap.Provider.registerIfAbsent(new WeakMapSuppliers.Guava());
     }
@@ -48,11 +48,11 @@ public class AgentTooling {
   private static final DDCachingPoolStrategy POOL_STRATEGY = new DDCachingPoolStrategy();
 
   public static <K, V> WeakCache<K, V> newWeakCache() {
-    return weakCacheProvider.newWeakCache();
+    return weakCacheProvider.<K, V>newWeakCache();
   }
 
   public static <K, V> WeakCache<K, V> newWeakCache(final long maxSize) {
-    return weakCacheProvider.newWeakCache(maxSize);
+    return weakCacheProvider.<K, V>newWeakCache(maxSize);
   }
 
   public static DDLocationStrategy locationStrategy() {
